Feature/implement basic functionality (#3)
* Add xml to dependency

* Add classes in core

* Add exceptions

* Add empty soap-request and -response classes

* Add empty plattform abstractions for sending requests

* Implement SoapRequest.

* Add XmlMissingElementException

* Make errorCode optional.

* Add overrides for toString

* Add documentation to exception-model

* Implement SoapResponse

* Add copyright-notice

* Fix typo in namespace

* Fix usage of wrong type

* Add ServiceEndpointConfiguration

Added to simplify passing configuration information. This also simplify addition of additional information when needed.

* Add const modifier to constructors

* Add HttpException

this error gets thrown if the status code from response of the service is not ok and does not contain a soap:Envelope with soap:Fault

* Rework exception hierarchy

All exceptions that are a consequence of a malformed response extend or implement the MalformedContentException instead.

* Fix import statements not using relativ sources

* Implement IoClient

* Add documentation

* Change return type for easier testing

* Fix faultCode being alway null

* Add tests for SoapRequest

* Add tests for SoapRequest

Previous commit contained tests for SoapResponse.

* Remove unnecessary statements

* Fix ressource leak

* Extract class into own file.

The file will hold all code that is needed to write tests.

* Add tests for IOClient

* Fix might return null error.

* Fix not skipping test which fails.

Test in question tests code that is mostly in a other library. This will be looked at later.